Email aggregate G4 (#217): rename EmailTemplate->Email, Zipper.Emails namespace
