+++ src/Emails/EmailCategory.cs
@@ -1,21 +1,20 @@
-namespace Zipper
+namespace Zipper.Emails;
+
+/// <summary>
+/// Email template categories for variety.
+/// </summary>
+public enum EmailCategory
 {
-    /// <summary>
-    /// Email template categories for variety.
-    /// </summary>
-    public enum EmailCategory
-    {
-        Business,
-        Personal,
-        Technical,
-        Marketing,
-        Legal,
-        Financial,
-        Notification,
-        Support,
-        Healthcare,
-        Education,
-        Ecommerce,
-        Travel,
-    }
+    Business,
+    Personal,
+    Technical,
+    Marketing,
+    Legal,
+    Financial,
+    Notification,
+    Support,
+    Healthcare,
+    Education,
+    Ecommerce,
+    Travel,
 }
