--- src/Zipper.Tests/EmailBuilderTests.cs
@@ -1,481 +1,0 @@
-using System.Text;
-using Xunit;
-
-namespace Zipper
-{
-    public class EmailBuilderTests
-    {
-        [Fact]
-        public void BuildEmail_BasicTemplate_CreatesValidEml()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Test Subject",
-                SentDate = new DateTime(2023, 1, 1, 12, 0, 0),
-                Body = "This is a test email body.",
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.True(result.Length > 0);
-
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("From: sender@example.com", content);
-            Assert.Contains("To: test@example.com", content);
-            Assert.Contains("Subject: Test Subject", content);
-            Assert.Contains("This is a test email body.", content);
-            Assert.Contains("text/plain; charset=utf-8", content);
-        }
-
-        [Fact]
-        public void BuildEmail_WithAttachment_CreatesMultipartContent()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Test with Attachment",
-                Body = "Email with attachment.",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = "test.txt",
-                Content = Encoding.UTF8.GetBytes("Test attachment content"),
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("multipart/mixed", content);
-            Assert.Contains("Content-Type: text/plain", content);
-            Assert.Contains("Content-Transfer-Encoding: base64", content);
-            Assert.Contains("Content-Disposition: attachment", content);
-            Assert.Contains("VGVzdCBhdHRhY2htZW50IGNvbnRlbnQ=", content); // Base64 encoded content
-        }
-
-        [Fact]
-        public void BuildEmail_WithCcAndBcc_IncludesHeaders()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "to@example.com",
-                From = "from@example.com",
-                Cc = "cc@example.com",
-                Bcc = "bcc@example.com",
-                Subject = "Test CC/BCC",
-                Body = "Test body",
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Cc: cc@example.com", content);
-            Assert.Contains("Bcc: bcc@example.com", content);
-        }
-
-        [Fact]
-        public void BuildEmail_HighPriority_IncludesPriorityHeaders()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "High Priority",
-                Body = "Important email",
-                IsHighPriority = true,
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("X-Priority: 1", content);
-            Assert.Contains("Priority: Urgent", content);
-        }
-
-        [Fact]
-        public void BuildEmail_WithReadReceipt_IncludesDispositionHeader()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Read Receipt Test",
-                Body = "Please confirm receipt",
-                RequestReadReceipt = true,
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Disposition-Notification-To: sender@example.com", content);
-        }
-
-        [Fact]
-        public void BuildEmail_WithReplyTo_IncludesReplyToHeader()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                ReplyTo = "replyto@example.com",
-                Subject = "Reply To Test",
-                Body = "Please reply to different address",
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Reply-To: replyto@example.com", content);
-        }
-
-        [Theory]
-        [InlineData("test.pdf", "application/pdf")]
-        [InlineData("document.docx", "application/vnd.openxmlformats-officedocument.wordprocessingml.document")]
-        [InlineData("spreadsheet.xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")]
-        [InlineData("image.jpg", "image/jpeg")]
-        [InlineData("image.png", "image/png")]
-        [InlineData("text.txt", "text/plain")]
-        [InlineData("unknown.xyz", "application/octet-stream")]
-        public void BuildEmail_DifferentFileTypes_CorrectContentType(string fileName, string expectedContentType)
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "File Type Test",
-                Body = "Test file type detection",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = fileName,
-                Content = new byte[] { 1, 2, 3, 4, 5 },
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains(expectedContentType, content);
-        }
-
-        [Fact]
-        public void BuildEmail_WithCustomContentType_UsesCustomType()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Custom Content Type",
-                Body = "Test custom content type",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = "custom.test",
-                Content = new byte[] { 1, 2, 3 },
-                ContentType = "application/custom-type",
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("application/custom-type", content);
-        }
-
-        [Fact]
-        public void BuildEmail_InlineAttachment_IncludesContentId()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Inline Attachment",
-                Body = "Test inline attachment",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = "image.png",
-                Content = new byte[] { 1, 2, 3 },
-                ContentId = "test-image",
-                IsInline = true,
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            // Assert
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Content-ID: <test-image>", content);
-            Assert.Contains("Content-Disposition: inline", content);
-        }
-
-        [Fact]
-        public void BuildEmail_LegacyInterface_CompatibleWithNew()
-        {
-            // Arrange
-            var to = "test@example.com";
-            var from = "sender@example.com";
-            var subject = "Legacy Test";
-            var sentDate = new DateTime(2023, 1, 1, 12, 0, 0);
-            var body = "Legacy interface test";
-            var attachment = ("test.txt", Encoding.UTF8.GetBytes("Test content"));
-
-            // Act
-            var legacyResult = EmailBuilder.BuildEmail(to, from, subject, sentDate, body, attachment);
-
-            var template = new EmailTemplate
-            {
-                To = to,
-                From = from,
-                Subject = subject,
-                SentDate = sentDate,
-                Body = body,
-            };
-            var attachmentInfo = new AttachmentInfo
-            {
-                FileName = attachment.Item1,
-                Content = attachment.Item2,
-            };
-            var newResult = EmailBuilder.BuildEmail(template, attachmentInfo);
-
-            // Assert - Both should produce valid EML content with same structure
-            Assert.NotNull(legacyResult);
-            Assert.NotNull(newResult);
-            Assert.True(legacyResult.Length > 0);
-            Assert.True(newResult.Length > 0);
-
-            // Both should contain the same key EML elements
-            var legacyContent = Encoding.UTF8.GetString(legacyResult);
-            var newContent = Encoding.UTF8.GetString(newResult);
-
-            Assert.Contains("From: sender@example.com", legacyContent);
-            Assert.Contains("To: test@example.com", legacyContent);
-            Assert.Contains("Subject: Legacy Test", legacyContent);
-
-            Assert.Contains("From: sender@example.com", newContent);
-            Assert.Contains("To: test@example.com", newContent);
-            Assert.Contains("Subject: Legacy Test", newContent);
-
-            // Both should have multipart content
-            Assert.Contains("multipart/mixed", legacyContent);
-            Assert.Contains("multipart/mixed", newContent);
-        }
-
-        [Fact]
-        public void BuildEmail_NullTemplate_ThrowsArgumentNullException()
-        {
-            // Act & Assert
-            Assert.Throws<ArgumentNullException>(() => EmailBuilder.BuildEmail(null!));
-        }
-
-        [Fact]
-        public void BuildEmail_LargeAttachment_HandlesCorrectly()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Large Attachment",
-                Body = "Email with large attachment",
-            };
-            var largeContent = new byte[1024 * 1024]; // 1MB
-            Random.Shared.NextBytes(largeContent);
-
-            var attachment = new AttachmentInfo
-            {
-                FileName = "large.dat",
-                Content = largeContent,
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            // Assert
-            Assert.True(result.Length > largeContent.Length);
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("base64", content.ToLowerInvariant());
-        }
-
-        [Fact]
-        public void BuildEmail_SpecialCharactersInHeaders_EncodesCorrectly()
-        {
-            // Arrange
-            var template = new EmailTemplate
-            {
-                To = "tëst@éxample.com",
-                From = "sênder@examplé.com",
-                Subject = "Tëst with spëcial chars: 你好",
-                Body = "Body with special characters: café, naïve",
-            };
-
-            // Act
-            var result = EmailBuilder.BuildEmail(template);
-
-            // Assert
-            Assert.NotNull(result);
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("tëst@éxample.com", content);
-            Assert.Contains("Tëst with spëcial chars: 你好", content);
-        }
-
-        [Fact]
-        public void BuildEmail_EmptyBodyAndSubject_CreatesValidEml()
-        {
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = string.Empty,
-                SentDate = new DateTime(2023, 1, 1),
-                Body = string.Empty,
-            };
-
-            var result = EmailBuilder.BuildEmail(template);
-
-            Assert.NotNull(result);
-            Assert.True(result.Length > 0);
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Subject:", content);
-        }
-
-        [Fact]
-        public void BuildEmail_DateTimeMinValue_HandlesGracefully()
-        {
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Date Test",
-                SentDate = DateTime.MinValue,
-                Body = "Test body",
-            };
-
-            var result = EmailBuilder.BuildEmail(template);
-
-            Assert.NotNull(result);
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("Date:", content);
-        }
-
-        [Fact]
-        public void BuildEmail_ZeroLengthAttachment_HandlesGracefully()
-        {
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Empty Attachment",
-                Body = "Test",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = "empty.txt",
-                Content = Array.Empty<byte>(),
-            };
-
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            Assert.NotNull(result);
-            var content = Encoding.UTF8.GetString(result);
-            Assert.Contains("multipart/mixed", content);
-        }
-
-        [Fact]
-        public void BuildEmail_NullAttachmentFileName_HandlesGracefully()
-        {
-            var template = new EmailTemplate
-            {
-                To = "test@example.com",
-                From = "sender@example.com",
-                Subject = "Null FileName",
-                Body = "Test",
-            };
-            var attachment = new AttachmentInfo
-            {
-                FileName = null!,
-                Content = new byte[] { 1, 2, 3 },
-            };
-
-            var result = EmailBuilder.BuildEmail(template, attachment);
-
-            Assert.NotNull(result);
-            Assert.True(result.Length > 0);
-        }
-
-        [Fact]
-        public void EmailBuilder_BackwardCompatibility_LegacyOverload()
-        {
-            // Arrange
-            var to = "test@example.com";
-            var from = "sender@example.com";
-            var subject = "Compatibility Test";
-            var sentDate = new DateTime(2023, 6, 15, 14, 30, 0);
-            var body = "Testing backward compatibility";
-            var attachment = ("test.txt", Encoding.UTF8.GetBytes("Attachment content"));
-
-            // Act - use the legacy overload of BuildEmail
-            var legacyResult = EmailBuilder.BuildEmail(to, from, subject, sentDate, body, attachment);
-
-            // Also build via the new template API
-            var template = new EmailTemplate
-            {
-                To = to,
-                From = from,
-                Subject = subject,
-                SentDate = sentDate,
-                Body = body,
-            };
-            var attachmentInfo = new AttachmentInfo
-            {
-                FileName = attachment.Item1,
-                Content = attachment.Item2,
-            };
-            var newResult = EmailBuilder.BuildEmail(template, attachmentInfo);
-
-            // Assert - Both should produce valid EML content
-            Assert.NotNull(legacyResult);
-            Assert.NotNull(newResult);
-            Assert.True(legacyResult.Length > 0);
-            Assert.True(newResult.Length > 0);
-
-            // Both should contain expected EML headers
-            var legacyContent = Encoding.UTF8.GetString(legacyResult);
-            var newContent = Encoding.UTF8.GetString(newResult);
-
-            Assert.Contains("From: sender@example.com", legacyContent);
-            Assert.Contains("To: test@example.com", legacyContent);
-            Assert.Contains("Subject: Compatibility Test", legacyContent);
-
-            Assert.Contains("From: sender@example.com", newContent);
-            Assert.Contains("To: test@example.com", newContent);
-            Assert.Contains("Subject: Compatibility Test", newContent);
-        }
-    }
-}
