--- src/Zipper.Tests/EmailTemplateSystemTests.cs
@@ -1,511 +1,0 @@
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Zipper
-{
-    public class EmailTemplateSystemTests
-    {
-        private readonly ITestOutputHelper output;
-
-        public EmailTemplateSystemTests(ITestOutputHelper output)
-        {
-            this.output = output;
-        }
-
-        [Fact]
-        public void GetRandomTemplate_BasicParameters_ReturnsValidEmailTemplate()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.False(string.IsNullOrEmpty(template.To));
-            Assert.False(string.IsNullOrEmpty(template.From));
-            Assert.False(string.IsNullOrEmpty(template.Subject));
-            Assert.False(string.IsNullOrEmpty(template.Body));
-            Assert.True(template.SentDate <= DateTime.Now);
-        }
-
-        [Theory]
-        [InlineData(EmailCategory.Business)]
-        [InlineData(EmailCategory.Personal)]
-        [InlineData(EmailCategory.Technical)]
-        [InlineData(EmailCategory.Marketing)]
-        [InlineData(EmailCategory.Legal)]
-        [InlineData(EmailCategory.Financial)]
-        [InlineData(EmailCategory.Notification)]
-        [InlineData(EmailCategory.Support)]
-        [InlineData(EmailCategory.Healthcare)]
-        [InlineData(EmailCategory.Education)]
-        [InlineData(EmailCategory.Ecommerce)]
-        [InlineData(EmailCategory.Travel)]
-        public void GetRandomTemplate_WithSpecificCategory_ReturnsCategorySpecificTemplates(EmailCategory category)
-        {
-            // Arrange
-            const int recipientIndex = 10;
-            const int senderIndex = 20;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, category);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.Contains($"recipient{recipientIndex:D3}@", template.To);
-            Assert.Contains($"sender{senderIndex:D3}@", template.From);
-
-            this.output.WriteLine($"Category: {category}, Subject: {template.Subject}");
-            this.output.WriteLine($"Body preview: {template.Body.Substring(0, Math.Min(100, template.Body.Length))}...");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_MultipleCalls_ReturnsVariedContent()
-        {
-            // Arrange
-            const int iterations = 50;
-            var subjects = new string[iterations];
-
-            // Act
-            for (int i = 0; i < iterations; i++)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2);
-                subjects[i] = template.Subject;
-            }
-
-            // Assert
-            var uniqueSubjects = subjects.Distinct().Count();
-            Assert.True(uniqueSubjects > 1, "Expected multiple different subjects");
-            Assert.True(uniqueSubjects >= iterations * 0.3, $"Expected at least 30% variety, got {uniqueSubjects}/{iterations}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_WithDifferentIndices_GeneratesUniqueEmailAddresses()
-        {
-            // Arrange
-            const int iterations = 20;
-            var emailAddresses = new (string to, string from)[iterations];
-
-            // Act
-            for (int i = 0; i < iterations; i++)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(i, i * 2);
-                emailAddresses[i] = (template.To, template.From);
-            }
-
-            // Assert
-            var uniqueToAddresses = emailAddresses.Select(e => e.to).Distinct().Count();
-            var uniqueFromAddresses = emailAddresses.Select(e => e.from).Distinct().Count();
-
-            Assert.Equal(iterations, uniqueToAddresses);
-            Assert.Equal(iterations, uniqueFromAddresses);
-        }
-
-        [Fact]
-        public void GetContextualTemplate_WithValidContext_ReturnsAppropriateTemplate()
-        {
-            // Arrange
-            var context = new EmailContext
-            {
-                RecipientIndex = 5,
-                SenderIndex = 10,
-                Category = EmailCategory.Business,
-                TemplateIndex = 0,
-                SentDate = new DateTime(2023, 6, 15),
-                IsHighPriority = true,
-                RequestReadReceipt = true,
-            };
-
-            // Act
-            var template = EmailTemplateSystem.GetContextualTemplate(context);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.Contains("recipient005@", template.To);
-            Assert.Contains("sender010@", template.From);
-            Assert.Equal(new DateTime(2023, 6, 15), template.SentDate);
-            Assert.True(template.IsHighPriority);
-            Assert.True(template.RequestReadReceipt);
-        }
-
-        [Fact]
-        public void GenerateEmailAddress_WithValidParameters_ReturnsRealisticEmailAddresses()
-        {
-            // Arrange
-            const int index = 42;
-            const string type = "user";
-
-            // Act
-            var emailAddress = EmailTemplateSystem.GenerateEmailAddress(index, type);
-
-            // Assert
-            Assert.Contains($"user{index:D3}@", emailAddress);
-            Assert.Contains("@", emailAddress);
-        }
-
-        [Theory]
-        [InlineData(0, "test")]
-        [InlineData(99, "admin")]
-        [InlineData(150, "user")]
-        [InlineData(1000, "manager")]
-        public void GenerateEmailAddress_WithDifferentIndicesAndTypes_ReturnsCorrectFormat(int index, string type)
-        {
-            // Act
-            var emailAddress = EmailTemplateSystem.GenerateEmailAddress(index, type);
-
-            // Assert
-            Assert.Contains($"{type}{index:D3}@", emailAddress);
-        }
-
-        [Fact]
-        public void GetRandomTemplate_BusinessCategory_ContainsBusinessContent()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, EmailCategory.Business);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.True(
-                template.Subject.Contains("Business") ||
-                template.Subject.Contains("Contract") ||
-                template.Subject.Contains("Meeting") ||
-                template.Subject.Contains("Partnership") ||
-                template.Subject.Contains("Budget") ||
-                template.Subject.Contains("Review") ||
-                template.Subject.Contains("Proposal"),
-                $"Expected business-related subject, got: {template.Subject}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_HealthcareCategory_ContainsHealthcareContent()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, EmailCategory.Healthcare);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.True(
-                template.Subject.Contains("Medical") ||
-                template.Subject.Contains("Appointment") ||
-                template.Subject.Contains("Lab") ||
-                template.Subject.Contains("Prescription") ||
-                template.Subject.Contains("Check-up") ||
-                template.Subject.Contains("Telehealth"),
-                $"Expected healthcare-related subject, got: {template.Subject}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_EducationCategory_ContainsEducationContent()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, EmailCategory.Education);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.True(
-                template.Subject.Contains("Course") ||
-                template.Subject.Contains("Assignment") ||
-                template.Subject.Contains("Progress") ||
-                template.Subject.Contains("Class") ||
-                template.Subject.Contains("Achievement"),
-                $"Expected education-related subject, got: {template.Subject}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_EcommerceCategory_ContainsEcommerceContent()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, EmailCategory.Ecommerce);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.True(
-                template.Subject.Contains("Order") ||
-                template.Subject.Contains("Shipped") ||
-                template.Subject.Contains("Offer") ||
-                template.Subject.Contains("Cart") ||
-                template.Subject.Contains("Review"),
-                $"Expected ecommerce-related subject, got: {template.Subject}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_TravelCategory_ContainsTravelContent()
-        {
-            // Arrange
-            const int recipientIndex = 1;
-            const int senderIndex = 2;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(recipientIndex, senderIndex, EmailCategory.Travel);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.True(
-                template.Subject.Contains("Booking") ||
-                template.Subject.Contains("Flight") ||
-                template.Subject.Contains("Hotel") ||
-                template.Subject.Contains("Insurance") ||
-                template.Subject.Contains("Rental"),
-                $"Expected travel-related subject, got: {template.Subject}");
-        }
-
-        [Fact]
-        public void GetRandomTemplate_NewCategories_HaveRealisticPlaceholdersReplaced()
-        {
-            // Arrange
-            var categories = new[]
-            {
-                EmailCategory.Healthcare,
-                EmailCategory.Education,
-                EmailCategory.Ecommerce,
-                EmailCategory.Travel,
-            };
-
-            // Placeholders that SHOULD be replaced in body
-            var bodyPlaceholders = new[]
-            {
-                "{recipient}", "{sender}", "{company}", "{department}", "{project}",
-                "{amount}", "{deadline}", "{meeting}", "{date}", "{place}", "{venue}", "{website}", "{service}",
-                "{reset_link}", "{quarter}", "{growth}", "{payment}", "{account}", "{start_time}", "{end_time}", "{month}",
-                "{gate}", "{seat}", "{rental_period}",
-            };
-
-            // Act & Assert
-            foreach (var category in categories)
-            {
-                for (int i = 0; i < 5; i++)
-                {
-                    var template = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2, category);
-
-                    // Check that known body placeholders are replaced
-                    foreach (var placeholder in bodyPlaceholders)
-                    {
-                        Assert.DoesNotContain(placeholder, template.Body);
-                    }
-
-                    this.output.WriteLine($"Category: {category}, Subject: {template.Subject}");
-                }
-            }
-        }
-
-        [Fact]
-        public void GetRandomTemplate_AllCategories_VarySentDateAppropriately()
-        {
-            // Arrange
-            var categories = Enum.GetValues<EmailCategory>();
-            var now = DateTime.Now;
-            var categoryDates = new System.Collections.Generic.Dictionary<EmailCategory, TimeSpan>();
-
-            // Act
-            foreach (var category in categories)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(1, 2, category);
-                var timeDifference = now - template.SentDate;
-                categoryDates[category] = timeDifference;
-            }
-
-            // Assert - Verify different categories have different date ranges
-            Assert.True(categoryDates.Count > 0);
-
-            // Travel emails should potentially have a wider range (older dates possible)
-            var travelTime = categoryDates[EmailCategory.Travel];
-            var supportTime = categoryDates[EmailCategory.Support];
-
-            this.output.WriteLine($"Travel time difference: {travelTime.TotalDays:F1} days");
-            this.output.WriteLine($"Support time difference: {supportTime.TotalDays:F1} days");
-        }
-
-        [Theory]
-        [InlineData(EmailCategory.Healthcare)]
-        [InlineData(EmailCategory.Education)]
-        [InlineData(EmailCategory.Ecommerce)]
-        [InlineData(EmailCategory.Travel)]
-        public void GetRandomTemplate_NewCategories_HaveAppropriateCcAndReplyToRates(EmailCategory category)
-        {
-            // Arrange
-            const int iterations = 100;
-            var templatesWithCc = 0;
-            var templatesWithReplyTo = 0;
-
-            // Act
-            for (int i = 0; i < iterations; i++)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2, category);
-                if (!string.IsNullOrEmpty(template.Cc))
-                {
-                    templatesWithCc++;
-                }
-
-                if (!string.IsNullOrEmpty(template.ReplyTo))
-                {
-                    templatesWithReplyTo++;
-                }
-            }
-
-            // Assert
-            var ccRate = (double)templatesWithCc / iterations;
-            var replyToRate = (double)templatesWithReplyTo / iterations;
-
-            this.output.WriteLine($"Category: {category}, CC rate: {ccRate:P2}, ReplyTo rate: {replyToRate:P2}");
-
-            // Should have some variation but not always present
-            Assert.InRange(ccRate, 0.0, 1.0);
-            Assert.InRange(replyToRate, 0.0, 1.0);
-        }
-
-        [Fact]
-        public void GetRandomTemplate_WithHighProbabilityCategories_HasMoreCcAddresses()
-        {
-            // Arrange
-            const int iterations = 100;
-            var businessCcCount = 0;
-            var marketingCcCount = 0;
-
-            // Act
-            for (int i = 0; i < iterations; i++)
-            {
-                var businessTemplate = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2, EmailCategory.Business);
-                var marketingTemplate = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2, EmailCategory.Marketing);
-
-                if (!string.IsNullOrEmpty(businessTemplate.Cc))
-                {
-                    businessCcCount++;
-                }
-
-                if (!string.IsNullOrEmpty(marketingTemplate.Cc))
-                {
-                    marketingCcCount++;
-                }
-            }
-
-            // Assert
-            Assert.True(
-                marketingCcCount > businessCcCount,
-                $"Expected marketing to have more CC addresses. Marketing: {marketingCcCount}, Business: {businessCcCount}");
-        }
-
-        [Fact]
-        public void EmailContext_RecordType_WorksCorrectly()
-        {
-            // Arrange
-            var context1 = new EmailContext
-            {
-                RecipientIndex = 1,
-                SenderIndex = 2,
-                Category = EmailCategory.Business,
-                TemplateIndex = 0,
-            };
-
-            // Act
-            var context2 = context1 with { };
-
-            // Assert
-            Assert.Equal(context1.RecipientIndex, context2.RecipientIndex);
-            Assert.Equal(context1.SenderIndex, context2.SenderIndex);
-            Assert.Equal(context1.Category, context2.Category);
-            Assert.Equal(context1.TemplateIndex, context2.TemplateIndex);
-            Assert.True(context1 == context2);
-        }
-
-        [Fact]
-        public void GetRandomTemplate_WithLargeIndices_HandlesCorrectly()
-        {
-            // Arrange
-            const int largeRecipientIndex = 999999;
-            const int largeSenderIndex = 888888;
-
-            // Act
-            var template = EmailTemplateSystem.GetRandomTemplate(largeRecipientIndex, largeSenderIndex);
-
-            // Assert
-            Assert.NotNull(template);
-            Assert.Contains($"recipient{largeRecipientIndex:D6}@", template.To);
-            Assert.Contains($"sender{largeSenderIndex:D6}@", template.From);
-        }
-
-        [Fact]
-        public void GetRandomTemplate_TemplateContent_HasRealisticFormatting()
-        {
-            // Arrange
-            const int iterations = 20;
-
-            // Act
-            for (int i = 0; i < iterations; i++)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2);
-
-                // Assert
-                Assert.False(string.IsNullOrWhiteSpace(template.Body));
-                Assert.Contains("\n", template.Body); // Should have line breaks
-
-                // Should have proper greeting and closing patterns
-                var bodyLower = template.Body.ToLowerInvariant();
-                var hasGreeting = bodyLower.Contains("dear") || bodyLower.Contains("hi") || bodyLower.Contains("hello") || bodyLower.Contains("welcome");
-                var hasClosing = bodyLower.Contains("regards") || bodyLower.Contains("sincerely") || bodyLower.Contains("best") || bodyLower.Contains("thank") || bodyLower.Contains("looking forward");
-
-                this.output.WriteLine($"Template {i + 1}: {template.Subject}");
-                this.output.WriteLine($"Has greeting: {hasGreeting}, Has closing: {hasClosing}");
-            }
-        }
-
-        [Fact]
-        public void GetRandomTemplate_BodyReplacements_AllPlaceholdersReplaced()
-        {
-            // Arrange
-            const int iterations = 50;
-
-            // These placeholders are replaced in BOTH subject and body
-            var commonPlaceholders = new[] { "{recipient}", "{sender}", "{date}", "{quarter}" };
-
-            // These placeholders are replaced only in body (not in subject generation)
-            var bodyOnlyPlaceholders = new[] { "{company}", "{department}", "{project}", "{amount}", "{deadline}", "{meeting}", "{place}", "{venue}", "{website}", "{service}", "{reset_link}", "{growth}", "{payment}", "{account}", "{start_time}", "{end_time}", "{month}" };
-
-            // These placeholders are replaced only in subject (not in body - this is a known implementation quirk)
-            var subjectOnlyPlaceholders = new[] { "{case}", "{invoice}", "{ticket}", "{course}" };
-
-            // Act & Assert
-            for (int i = 0; i < iterations; i++)
-            {
-                var template = EmailTemplateSystem.GetRandomTemplate(i + 1, i + 2);
-
-                // Check common placeholders are replaced in both
-                foreach (var placeholder in commonPlaceholders)
-                {
-                    Assert.DoesNotContain(placeholder, template.Subject);
-                    Assert.DoesNotContain(placeholder, template.Body);
-                }
-
-                // Check subject-only placeholders are replaced in subject only
-                foreach (var placeholder in subjectOnlyPlaceholders)
-                {
-                    Assert.DoesNotContain(placeholder, template.Subject);
-                }
-
-                // Check body-only placeholders are replaced in body
-                foreach (var placeholder in bodyOnlyPlaceholders)
-                {
-                    Assert.DoesNotContain(placeholder, template.Body);
-                }
-            }
-        }
-    }
-}
